+++ x/finality/types/expected_keepers.go
@@ -36,6 +36,6 @@ type CheckpointingKeeper interface {
 type IncentiveKeeper interface {
 	RewardBTCStaking(ctx context.Context, height uint64, filteredDc *VotingPowerDistCache, voters map[string]struct{})
 	IndexRefundableMsg(ctx context.Context, msg sdk.Msg)
-	BtcDelegationActivated(ctx context.Context, fp, del sdk.AccAddress, sat uint64) error
-	BtcDelegationUnbonded(ctx context.Context, fp, del sdk.AccAddress, sat uint64) error
+	AddEventBtcDelegationActivated(ctx context.Context, height uint64, fp, del sdk.AccAddress, sat uint64) error
+	AddEventBtcDelegationUnbonded(ctx context.Context, height uint64, fp, del sdk.AccAddress, sat uint64) error
 }