+++ x/finality/types/mocked_keepers.go
@@ -307,32 +307,32 @@ func (m *MockIncentiveKeeper) EXPECT() *MockIncentiveKeeperMockRecorder {
 	return m.recorder
 }
 
-// BtcDelegationActivated mocks base method.
-func (m *MockIncentiveKeeper) BtcDelegationActivated(ctx context.Context, fp, del types1.AccAddress, sat uint64) error {
+// AddEventBtcDelegationActivated mocks base method.
+func (m *MockIncentiveKeeper) AddEventBtcDelegationActivated(ctx context.Context, height uint64, fp, del types1.AccAddress, sat uint64) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "BtcDelegationActivated", ctx, fp, del, sat)
+	ret := m.ctrl.Call(m, "AddEventBtcDelegationActivated", ctx, height, fp, del, sat)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
-// BtcDelegationActivated indicates an expected call of BtcDelegationActivated.
-func (mr *MockIncentiveKeeperMockRecorder) BtcDelegationActivated(ctx, fp, del, sat interface{}) *gomock.Call {
+// AddEventBtcDelegationActivated indicates an expected call of AddEventBtcDelegationActivated.
+func (mr *MockIncentiveKeeperMockRecorder) AddEventBtcDelegationActivated(ctx, height, fp, del, sat interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "BtcDelegationActivated", reflect.TypeOf((*MockIncentiveKeeper)(nil).BtcDelegationActivated), ctx, fp, del, sat)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddEventBtcDelegationActivated", reflect.TypeOf((*MockIncentiveKeeper)(nil).AddEventBtcDelegationActivated), ctx, height, fp, del, sat)
 }
 
-// BtcDelegationUnbonded mocks base method.
-func (m *MockIncentiveKeeper) BtcDelegationUnbonded(ctx context.Context, fp, del types1.AccAddress, sat uint64) error {
+// AddEventBtcDelegationUnbonded mocks base method.
+func (m *MockIncentiveKeeper) AddEventBtcDelegationUnbonded(ctx context.Context, height uint64, fp, del types1.AccAddress, sat uint64) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "BtcDelegationUnbonded", ctx, fp, del, sat)
+	ret := m.ctrl.Call(m, "AddEventBtcDelegationUnbonded", ctx, height, fp, del, sat)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
-// BtcDelegationUnbonded indicates an expected call of BtcDelegationUnbonded.
-func (mr *MockIncentiveKeeperMockRecorder) BtcDelegationUnbonded(ctx, fp, del, sat interface{}) *gomock.Call {
+// AddEventBtcDelegationUnbonded indicates an expected call of AddEventBtcDelegationUnbonded.
+func (mr *MockIncentiveKeeperMockRecorder) AddEventBtcDelegationUnbonded(ctx, height, fp, del, sat interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "BtcDelegationUnbonded", reflect.TypeOf((*MockIncentiveKeeper)(nil).BtcDelegationUnbonded), ctx, fp, del, sat)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddEventBtcDelegationUnbonded", reflect.TypeOf((*MockIncentiveKeeper)(nil).AddEventBtcDelegationUnbonded), ctx, height, fp, del, sat)
 }
 
 // IndexRefundableMsg mocks base method.